# openapi/vault-admin-api.yaml
openapi: 3.0.3
info:
  title: Vault Admin API
  description: |
    **Administrative API for Vault** - User and system management for administrators.
    
    This API is designed for admin dashboards and management tools. It provides:
    - Full user management (CRUD, suspend, audit)
    - Organization management (as admin)
    - Tenant configuration
    - Audit logs access
    - System settings
    
    ## Authentication
    
    All endpoints require:
    1. JWT Bearer token (same as Client API)
    2. Admin role (`admin` or `owner`)
    
    ```
    Authorization: Bearer <jwt-token>
    ```
    
    ## Authorization
    
    Endpoints have role-based access:
    - `owner` - Full access to everything
    - `admin` - Can manage users and orgs, cannot delete tenant
    
    ## Multi-Tenancy
    
    Admin requests are scoped to a tenant. Admins can only manage resources within their tenant.
  version: 0.1.0
  contact:
    name: Vault Team
  license:
    name: MIT OR Apache-2.0
servers:
  - url: https://api.vault.dev/api/v1/admin
    description: Production
  - url: http://localhost:3000/api/v1/admin
    description: Local development
...
paths:
...
  /users:
    get:
      summary: List all users
      description: |
        Get paginated list of all users in the tenant.
        Supports filtering by status, email, etc.
      operationId: listUsers
      tags: [User Management]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [active, pending, suspended, deactivated]
        - name: email
          in: query
          schema:
            type: string
          description: Filter by email (partial match)
        - name: org_id
          in: query
          schema:
            type: string
          description: Filter by organization membership
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUserResponse'
                  total:
                    type: integer
                  page:
                    type: integer
                  per_page:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create user (admin)
      description: Create a new user account as admin
      operationId: createUser
      tags: [User Management]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{userId}:
    get:
      summary: Get user details
      operationId: getUser
      tags: [User Management]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update user
      operationId: updateUser
      tags: [User Management]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete user (hard delete)
      description: Permanently delete a user account
      operationId: deleteUser
      tags: [User Management]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /scim/v2/Users:
    get:
      summary: List SCIM users
      operationId: scimListUsers
      tags: [SCIM]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: SCIM users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScimListResponse'
              example:
                schemas: ['urn:ietf:params:scim:api:messages:2.0:ListResponse']
                totalResults: 1
                startIndex: 1
                itemsPerPage: 20
                Resources:
                  - id: scim_user_123
                    userName: user@acme.com
                    active: true
                    emails:
                      - value: user@acme.com
                        primary: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Create SCIM user
      operationId: scimCreateUser
      tags: [SCIM]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScimUser'
            example:
              userName: user@acme.com
              active: true
              emails:
                - value: user@acme.com
                  primary: true
      responses:
        '201':
          description: SCIM user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScimUser'
              example:
                id: scim_user_123
                userName: user@acme.com
                active: true
                emails:
                  - value: user@acme.com
                    primary: true
        '400':
          $ref: '#/components/responses/BadRequest'
  /scim/v2/Users/{userId}:
    get:
      summary: Get SCIM user
      operationId: scimGetUser
      tags: [SCIM]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: SCIM user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScimUser'
              example:
                id: scim_user_123
                userName: user@acme.com
                active: true
                emails:
                  - value: user@acme.com
                    primary: true
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update SCIM user
      operationId: scimPatchUser
      tags: [SCIM]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScimPatchRequest'
            example:
              schemas: ['urn:ietf:params:scim:api:messages:2.0:PatchOp']
              Operations:
                - op: replace
                  path: active
                  value: false
      responses:
        '200':
          description: SCIM user updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScimUser'
              example:
                id: scim_user_123
                userName: user@acme.com
                active: false
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      summary: Delete SCIM user
      operationId: scimDeleteUser
      tags: [SCIM]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: SCIM user deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /scim/v2/Groups:
    get:
      summary: List SCIM groups
      operationId: scimListGroups
      tags: [SCIM]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: SCIM groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScimListResponse'
              example:
                schemas: ['urn:ietf:params:scim:api:messages:2.0:ListResponse']
                totalResults: 1
                startIndex: 1
                itemsPerPage: 20
                Resources:
                  - id: scim_group_123
                    displayName: Engineering
                    members: []
    post:
      summary: Create SCIM group
      operationId: scimCreateGroup
      tags: [SCIM]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScimGroup'
            example:
              displayName: Engineering
              members: []
      responses:
        '201':
          description: SCIM group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScimGroup'
              example:
                id: scim_group_123
                displayName: Engineering
                members: []
  /scim/v2/Groups/{groupId}:
    get:
      summary: Get SCIM group
      operationId: scimGetGroup
      tags: [SCIM]
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: SCIM group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScimGroup'
              example:
                id: scim_group_123
                displayName: Engineering
                members: []
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update SCIM group
      operationId: scimPatchGroup
      tags: [SCIM]
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScimPatchRequest'
            example:
              schemas: ['urn:ietf:params:scim:api:messages:2.0:PatchOp']
              Operations:
                - op: replace
                  path: displayName
                  value: Engineering
      responses:
        '200':
          description: SCIM group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScimGroup'
              example:
                id: scim_group_123
                displayName: Engineering
                members: []
    delete:
      summary: Delete SCIM group
      operationId: scimDeleteGroup
      tags: [SCIM]
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: SCIM group deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ============ DIRECTORY ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Admin permission required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }
    MessageResponse:
      type: object
      properties:
        message: { type: string }
...
    CreateUserRequest:
      type: object
      required: [email, name]
      properties:
        email: { type: string, format: email }
        name: { type: string }
        emailVerified: { type: boolean }
    UpdateUserRequest:
      type: object
      properties:
        email: { type: string, format: email }
        name: { type: string }
        status: { type: string, enum: [active, pending, suspended, deactivated] }
    AdminUserResponse:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        email_verified: { type: boolean }
        name: { type: string, nullable: true }
        status: { type: string }
        last_login_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        organization_count: { type: integer }
...
    ScimListResponse:
      type: object
      properties:
        schemas:
          type: array
          items: { type: string }
        totalResults: { type: integer }
        startIndex: { type: integer }
        itemsPerPage: { type: integer }
        Resources:
          type: array
          items: { type: object }
      example:
        schemas: ['urn:ietf:params:scim:api:messages:2.0:ListResponse']
        totalResults: 0
        startIndex: 1
        itemsPerPage: 20
        Resources: []
    ScimUser:
      type: object
      properties:
        id: { type: string }
        userName: { type: string }
        name: { type: object }
        active: { type: boolean }
        emails:
          type: array
          items: { type: object }
        externalId: { type: string }
        meta: { type: object }
      example:
        id: scim_user_123
        userName: user@acme.com
        name:
          givenName: Jane
          familyName: Doe
        active: true
        emails:
          - value: user@acme.com
            primary: true
        externalId: ext_123
        meta:
          resourceType: User
    ScimGroup:
      type: object
      properties:
        id: { type: string }
        displayName: { type: string }
        members:
          type: array
          items: { type: object }
        meta: { type: object }
      example:
        id: scim_group_123
        displayName: Engineering
        members: []
        meta:
          resourceType: Group
    ScimPatchRequest:
      type: object
      properties:
        schemas:
          type: array
          items: { type: string }
        Operations:
          type: array
          items: { type: object }
      example:
        schemas: ['urn:ietf:params:scim:api:messages:2.0:PatchOp']
        Operations:
          - op: replace
            path: active
            value: false

    # Directory
...